Update minidomparser.py
In order to parse an XML document using minidom, we must first import it from the xml.dom module. This module uses the parse function to create a DOM object. The parse function has the following syntax:
xml.dom.minidom.parse(filename_or_file[, parser[, bufsize]])
where the file name can be a string containing the file name or a file-type object. The function returns a document, which can be handled as an XML type. Thus, we can use the function getElementByTagName() to find a specific tag(s). 
As each node can be treated as an object, we can access the attributes and text of an element using the properties of the object. In this example, we have accessed the attributes and text of a specific node, and of all nodes together.